add theme support helper
