+++ src/ShopifyAppProvider.php
@@ -39,6 +39,7 @@ use Osiset\ShopifyApp\Messaging\Jobs\ScripttagInstaller;
 use Osiset\ShopifyApp\Messaging\Jobs\WebhookInstaller;
 use Osiset\ShopifyApp\Services\ApiHelper;
 use Osiset\ShopifyApp\Services\ChargeHelper;
+use Osiset\ShopifyApp\Services\ThemeHelper;
 use Osiset\ShopifyApp\Storage\Commands\Charge as ChargeCommand;
 use Osiset\ShopifyApp\Storage\Commands\Shop as ShopCommand;
 use Osiset\ShopifyApp\Storage\Observers\Shop as ShopObserver;
@@ -177,7 +178,7 @@ class ShopifyAppProvider extends ServiceProvider
         $this->app->bind(VerifyThemeSupportAction::class, function ($app) {
             return new VerifyThemeSupportAction(
                 $app->make(IShopQuery::class),
-                $app->make(IShopCommand::class)
+                $app->make(ThemeHelper::class)
             );
         });
